refactor DBConnectionManager to trait and create PostgresConnectionManager
