+++ server/src/main.rs
@@ -9,6 +9,7 @@ use tower_http::trace::TraceLayer;
 use crate::config::Config;
 use crate::config::Context;
 use crate::database::cache::CacheManager;
+use crate::database::postgres::PostgresConnectionManager;
 use crate::database::DBConnectionManager;
 use crate::prelude::*;
 use crate::routes::Build;
@@ -39,7 +40,8 @@ async fn main() -> Result<()> {
         "Initializing database connection pools..."
     );
 
-    let conn_manager = DBConnectionManager::get_connections(&config);
+    let conn_manager: Box<dyn DBConnectionManager> =
+        Box::new(PostgresConnectionManager::get_connections(&config));
 
     let router = axum::Router::build().layer(
         ServiceBuilder::new()
