+++ server/src/routes/info.rs
@@ -1,8 +1,6 @@
 use axum::{http::StatusCode, response::IntoResponse, routing::get, Extension, Json, Router};
 use serde::{Deserialize, Serialize};
 
-use crate::database::archive::fetch_chain_tip;
-use crate::database::archive::fetch_latest_slot;
 use crate::prelude::*;
 
 pub(crate) fn router() -> Router {
@@ -17,8 +15,8 @@ struct GetCoreApiInfoResponse {
 
 #[allow(clippy::unused_async)]
 async fn get_core_api_info(ctx: Extension<crate::Context>) -> Result<impl IntoResponse> {
-    let chain_tip = fetch_chain_tip(&ctx.conn_manager)?;
-    let current_slot = fetch_latest_slot(&ctx.conn_manager)?;
+    let chain_tip = ctx.conn_manager.fetch_chain_tip()?;
+    let current_slot = ctx.conn_manager.fetch_latest_slot()?;
 
     let response = GetCoreApiInfoResponse {
         chain_tip,
